split payment functionality done It's not perfect but we can clean it up later specifically the reciept generating
